added extended logging for stocks and music

+++ commands/musicPlay.js
@@ -1,7 +1,8 @@
 const Discord = require('discord.js');
 const fs = require('fs');
 const ytdl = require('ytdl-core');
-const ytSearch = require( 'yt-search' );
+const ytSearch = require('yt-search');
+const winston = require('winston');
 const SQLite = require("better-sqlite3");
 const serverVolumeSql = new SQLite('./db/serverVolume.sqlite');
 const songQueue = new SQLite('./db/songQueue.sqlite');
@@ -33,30 +34,36 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
             return;
         }
 
+        // const dancingDaddy = "https://media.giphy.com/media/8Bl2dceNUgY7sylmiU/giphy.gif"
         const embed = new Discord.RichEmbed();
         embed
             .setColor(alexaColor)
-            .setAuthor(`Let's get jiggy with it, ${message.author.username}`)
-            .setThumbnail(`https://media.giphy.com/media/8Bl2dceNUgY7sylmiU/giphy.gif`)
-            .setImage(`https://i.ytimg.com/vi/${video.videoId}/mqdefault.jpg`)
-            .addField(`${video.name}`,`https://www.youtube.com/watch?v=${video.videoId}`)
+            .setAuthor(`Let's get jiggy with it, my dudes`)
+            .setThumbnail(`https://i.ytimg.com/vi/${video.videoId}/mqdefault.jpg`)
+            .addField(`${video.name}`, `https://www.youtube.com/watch?v=${video.videoId}`)
             .setFooter(`If Alexa connected to the voice channel but isn't outputting any sound, just tell her to STFU and then try again. This is a known bug.`)
-        
-            message.channel.send(embed)
+
+        message.channel.send(embed)
             .then(newMessage => {
                 newMessage.react('▶')
-                    .then(() => {newMessage.react('⏸')
-                        .then(() => {newMessage.react('⏭')
-                            .then(() => {newMessage.react('⏹')})
-                        })
+                    .then(() => {
+                        newMessage.react('⏸')
+                            .then(() => {
+                                newMessage.react('⏭')
+                                    .then(() => {
+                                        newMessage.react('⏹')
+                                    })
+                            })
                     });
                 const filter = (reaction, user) => {
-                    return reaction.emoji.name === '▶' && user.id !== client.user.id || 
-                    reaction.emoji.name === '⏸' && user.id !== client.user.id || 
-                    reaction.emoji.name === '⏭' && user.id !== client.user.id || 
-                    reaction.emoji.name === '⏹' && user.id !== client.user.id;
+                    return reaction.emoji.name === '▶' && user.id !== client.user.id ||
+                        reaction.emoji.name === '⏸' && user.id !== client.user.id ||
+                        reaction.emoji.name === '⏭' && user.id !== client.user.id ||
+                        reaction.emoji.name === '⏹' && user.id !== client.user.id;
                 }
-                let emojiPause = new Discord.ReactionCollector(newMessage, filter, {time: 600000})
+                let emojiPause = new Discord.ReactionCollector(newMessage, filter, {
+                    time: 600000
+                })
 
                 emojiPause.on('collect', (reaction, reactionCollector) => {
                     if (reaction.emoji.name === '⏸') {
@@ -78,16 +85,44 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
                     }
                 })
             })
-        console.log("playing:");
-        console.log(video);
+
+        const musicLog = winston.createLogger({
+            level: 'info',
+            format: winston.format.combine(
+                winston.format.timestamp({
+                    format: 'YYYY-MM-DD HH:mm:ss'
+                }),
+                winston.format.json()
+            ),
+            transports: [
+                new winston.transports.File({
+                    filename: './logs/alexaMusic.log'
+                })
+            ]
+        })
+
+        musicLog.log({
+            level: 'info',
+            guildName: message.guild.name,
+            username: [message.author.username],
+            videoName: video.name,
+            videoId: video.videoId,
+            playOrStop: 'play',
+            endReason: null
+        })
+
         const channel = message.member.voiceChannel;
         const getServerVolume = serverVolumeSql.prepare("SELECT * FROM serverVolume WHERE guildId = ?").get(message.guild.id);
         if (!getServerVolume) {
-            streamOptions = {volume: 0.5};
+            streamOptions = {
+                volume: 0.5
+            };
         } else {
-            streamOptions = {volume: getServerVolume.volume};
+            streamOptions = {
+                volume: getServerVolume.volume
+            };
         }
-        
+
         if (!message.guild.voiceConnection) {
             let currentlyPlaying = require('../currentlyPlaying.json');
             currentlyPlaying.total++;
@@ -102,19 +137,49 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
 
                 let ytdlOptions = {};
                 if (video.seconds === 0) {
-                    ytdlOptions = {quality: 'highestaudio'};
+                    ytdlOptions = {
+                        quality: 'highestaudio'
+                    };
                     console.log(ytdlOptions);
                     console.log(video.seconds);
                 } else {
-                    ytdlOptions = {quality: 'highestaudio'};
+                    ytdlOptions = {
+                        quality: 'highestaudio'
+                    };
                     console.log(ytdlOptions);
                     console.log(video.seconds);
                 }
                 const stream = ytdl(`https://www.youtube.com/watch?v=${video.videoId}`, ytdlOptions)
                 const dispatcher = connection.playStream(stream, streamOptions);
 
+                const checkIfAlone = setInterval(() => {
+                    const notAlone = dispatcher.player.voiceConnection.channel.members.array()[1];
+                    if (!notAlone) {
+                        endReason = "alone";
+                        message.channel.send("Nobody is in the voice channel anymore, so I'm just gonna... make.. my.. way.. out of here...")
+                        connection.disconnect();
+                        return;
+                    }
+                }, 5000)
+
                 dispatcher.on("end", () => {
-                    console.log("dispatcher ended")
+                    const fullUserArray = dispatcher.player.voiceConnection.channel.members.array();
+                    const usernameArray = [];
+                    fullUserArray.forEach((element) => {
+                        usernameArray.push(element.user.username);
+                    })
+
+                    musicLog.log({
+                        level: 'info',
+                        guildName: dispatcher.player.voiceConnection.channel.guild.name,
+                        username: usernameArray,
+                        videoName: null,
+                        videoId: null,
+                        playOrStop: 'stop',
+                        endReason: endReason
+                    })
+
+                    clearInterval(checkIfAlone);
 
                     if (endReason === "stfu") {
                         endReason = "none";
@@ -135,11 +200,17 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
                         return;
                     }
 
+                    if (endReason === "alone") {
+                        endReason = "none";
+                        console.log("alone is the reason");
+                        return;
+                    }
+
                     const getServerQueue = songQueue.prepare("SELECT * FROM songQueue WHERE guildId = ? ORDER BY sortOrder ASC").get(message.guild.id);
                     console.log(getServerQueue);
                     if (getServerQueue) {
                         console.log("should be playing the next song")
-                        musicPlay(message,`alexa play fromalexaqueue https://youtube.com/watch?v=${getServerQueue.videoId}`, `alexa play fromalexaqueue https://youtube.com/watch?v=${getServerQueue.videoId}`, client);
+                        musicPlay(message, `alexa play fromalexaqueue https://youtube.com/watch?v=${getServerQueue.videoId}`, `alexa play fromalexaqueue https://youtube.com/watch?v=${getServerQueue.videoId}`, client);
                         songQueue.prepare("DELETE FROM songQueue WHERE guildId = ? AND videoId = ?").run(message.guild.id, getServerQueue.videoId);
                         endReason = "none";
                         return;
@@ -161,7 +232,7 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
                     }
                 })
             })
-            .catch( (error) => {
+            .catch((error) => {
                 if (error) {
                     setTimeout(() => {
                         message.channel.send(`Something went wrong! Alexa is sorry, bb. Try again or try a different search term.`)
@@ -192,7 +263,7 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
         // PLAY BY USING A DIRECT LINK TO A YOUTUBE PLAYLIST, WHICH ADDS THE PLAYLIST TO THE SERVER QUEUE AND PLAYS THE VIDEO ID IN THE URL
         //
         if (msgContent.includes("list=")) {
-            
+
             musicQueue(message);
 
             if (msgContent.includes("?v=")) {
@@ -201,7 +272,7 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
                 setTimeout(() => {
                     songQueue.prepare("DELETE FROM songQueue WHERE guildId = ? AND videoId = ?").run(message.guild.id, caseSensitiveContent.split("&list=")[0].split("?v=")[1])
                 }, 2000)
-                
+
                 return;
             } else {
                 return;
@@ -245,11 +316,13 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
                     name: result.player_response.videoDetails.title,
                     seconds: parseInt(result.player_response.videoDetails.lengthSeconds)
                 }
-                
+
                 if ((caseSensitiveContent.includes("fromalexaqueue")) === false) {
                     if (message.guild.voiceConnection) {
                         endReason = "play"
-                    } else {endReason = "none"}
+                    } else {
+                        endReason = "none"
+                    }
                 }
 
                 playThis(message, videoObj);
@@ -267,7 +340,7 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
                 message.channel.send(`Something went wrong! Alexa is sorry, bb. Try again or try a different search term.`)
                 return;
             }
-            
+
             const videos = result.videos
             var firstResult = videos[0]
             let videoObj;
@@ -286,11 +359,13 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
                 message.channel.send("That video doesn't seem to have a valid video ID. If you think this message is an error, please let me know on the Alexa Discord server: https://discord.gg/PysGrtD")
                 return;
             }
-            
+
             if ((caseSensitiveContent.includes("fromalexaqueue")) === false) {
                 if (message.guild.voiceConnection) {
                     endReason = "play"
-                } else {endReason = "none"}
+                } else {
+                    endReason = "none"
+                }
             }
 
             playThis(message, videoObj);
@@ -302,4 +377,8 @@ async function musicPlay(message, msgContent, caseSensitiveContent, client) {
     }
 }
 
-module.exports = { getEndReason, setEndReason, musicPlay };
+module.exports = {
+    getEndReason,
+    setEndReason,
+    musicPlay
+};
